requirements.txt
Django==4.2
mongoengine==0.24.2
djangorestframework==3.14.0
fastapi
uvicorn
uvicorn[standard]